[第7組] Restructure and Comments (#3)
* refactor: comments, fixed: constants

* refactor: comments, fixed: constants

* refactor: refactored structure to follow FastAPI bigger application structure

* refactor: refactored structure to follow FastAPI bigger application structure

* feat: document path

* feat: document path

* fixed: pydantic-settings

* feat: function comments

* feat: document path

* refactor: refactored structure to follow FastAPI bigger application structure
