+++ backend/dockerfile
@@ -2,7 +2,7 @@
 FROM python:3.10-slim
 
 # 設置工作目錄
-WORKDIR /app
+WORKDIR /backend
 
 # 將requirements.txt複製到容器内的工作目錄
 COPY requirements.txt .
@@ -17,8 +17,8 @@ COPY . .
 EXPOSE 8000
 
 # 定義環境變量
-ENV MODULE_NAME="app.main"
-ENV VARIABLE_NAME="app"
+ENV MODULE_NAME="backend.src.main"
+ENV VARIABLE_NAME="backend"
 
 # 使用 uvicorn 運行 fastapi 應用
-CMD ["uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "8000"]
+CMD ["uvicorn", "app.src.main:app", "--host", "0.0.0.0", "--port", "8000"]
