+++ backend/src/README.md
@@ -1,0 +1,23 @@
+## document path
+
+src  
+│  config.py  
+│  database.py  
+│  main.py  
+│  
+├─ auth  
+│  ├─ models.py  
+│  ├─ service.py  
+│  
+├─ news  
+│  ├─ models.py  
+│  ├─ router.py  
+│  ├─ schemas.py  
+│  ├─ service.py  
+│  
+├─ pricing  
+│  ├─ router.py  
+│  
+└─ users  
+│  ├─ router.py  
+│  ├─ schemas.py
